+++ CHANGELOG.rst
@@ -144,6 +144,8 @@ Added
   `#259 <https://github.com/openego/eGon-data/issues/259>`_
 * Distribute wind offshore capacities
   `#329 <https://github.com/openego/eGon-data/issues/329>`_
+* Assign weather cell id to weather dependant power plants
+  `#330 <https://github.com/openego/eGon-data/issues/330>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159